+++ Sources/Domain/GoogleDrive/Models/Email.swift
@@ -1,0 +1,11 @@
+//
+//  Email.swift
+//  Domain_GoogleDrive
+//
+//  Created by Jaewon Yun on 5/22/24.
+//  Copyright © 2024 woin2ee. All rights reserved.
+//
+
+import Foundation
+
+public typealias Email = String