+++ Sources/IOSScene/UserSettings/UserSettings/UserSettingsDataSource/UserSettingsItemModel.swift
@@ -1,14 +1,9 @@
-//
-//  UserSettingsItemModel.swift
-//  UserSettings
-//
-//  Created by Jaewon Yun on 11/28/23.
-//  Copyright © 2023 woin2ee. All rights reserved.
-//
-
 import IOSSupport
 
-enum UserSettingsItemModel {
-    case disclosureIndicator(DisclosureIndicatorCell.Model)
-    case button(ButtonCell.Model)
+extension UserSettingsViewController {
+
+    enum ItemModel {
+        case disclosureIndicator(DisclosureIndicatorCell.Model)
+        case button(ButtonCell.Model)
+    }
 }